requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-crispy-forms==1.11.2
django-filter==2.4.0
django-cors-headers==3.7.0
python-dotenv==0.19.1
Pillow
celery
redis==3.5.3
xlwt==1.3.0
colorama==0.4.6
setuptools
dj_rest_auth

